+++ BlazorFrontend/Utility Class/AppStateService.cs
@@ -2,5 +2,5 @@ namespace BlazorFrontend.Utility_Class;
 
 public class AppStateService
 {
-    public float? Cambio { get; set; }
+    public decimal? Cambio { get; set; }
 }
