+++ Modelos/Models/Dtos/EmpresaMonedaDto.cs
@@ -11,7 +11,7 @@ public sealed class EmpresaMonedaDto
     }
     public int IdEmpresaMoneda { get; set; }
 
-    public float? Cambio { get; set; }
+    public decimal? Cambio { get; set; }
 
     public EstadoEmpresaMoneda Estado { get; set; } = EstadoEmpresaMoneda.Abierto;
