+++ Modelos/Models/EmpresaMoneda.cs
@@ -14,7 +14,7 @@ public sealed class EmpresaMoneda
     [Key]
     public int IdEmpresaMoneda { get; set; }
 
-    public float? Cambio { get; set; }
+    public decimal? Cambio { get; set; }
 
     public EstadoEmpresaMoneda Estado { get; set; } = EstadoEmpresaMoneda.Abierto;
